// A0721I1-Repo-FE/coffee-house/src/app/model/product.ts
import {TypeProduct} from './typeProduct';

export interface Product {
  idProduct?: number;
  codeProduct?: string;
  nameProduct?: string;
  quatityProduct?: number;
  priceProduct?: number;
  imageProduct?: string;
  descriptionProduct?: string;
  createAt?: string;
  amountProduct?: number;
  typeProduct?: any;
}